HackerNews/app/src/main/java/propertyguru/hackernews/model/Comment.java
package propertyguru.hackernews.model;

import java.util.List;

/**
 * Created by hitesh on 6/23/17.
 */

public class Comment {

    public String text;
    public long time;
    public String by;
    public long id;
    public String type;
    public List<Long> kids;
    public int depth = 0;
    public String getText() {
        return text;
    }

    public void setText(String text) {
        this.text = text;
    }

    public long getTime() {
        return time;
    }

    public void setTime(Long time) {
        this.time = time;
    }

    public String getBy() {
        return by;
    }

    public void setBy(String by) {
        this.by = by;
    }

    public long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public List<Long> getKids() {
        return kids;
    }

    public void setKids(List<Long> kids) {
        this.kids = kids;
    }

    public int getDepth() {
        return depth;
    }

    public void setDepth(int depth) {
        this.depth = depth;
    }


}
